requirements.txt
docopt
flake8
jupyter
mne
mne-features
networkx
pandasql
pylint
python-dotenv
PyYAML
pep8-naming
scikit-learn
scikit-posthocs
scipy
statsmodels